+++ src/containerapp/HISTORY.rst
@@ -5,7 +5,7 @@ Release History
 upcoming
 ++++++
 * 'az containerapp auth update': Support authenticating blob storage token store using managed identity with `--blob-container-uri` and `--blob-container-identity`.
-* 'az containerapp up --model-registry --model-name --model-version: Support deploying Foundry model to Container App.'
+* 'az containerapp up --model-registry --model-name --model-version: Support deploying Azure AI Foundry model to Container App.'
 
 1.1.0b4
 ++++++